% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcPercentMissing.R
\name{calcPercentMissing}
\alias{calcPercentMissing}
\title{Computes the percentage of missing items in a subscale for each subject.}
\usage{
calcPercentMissing(dataFrame, items, idxOfSubj = 1)
}
\arguments{
\item{dataFrame}{A data frame containing the appropriate measure items.}

\item{items}{A vector of the items for which missingness is to be computed.}

\item{idxOfSubj}{A numeric index giving the position in the original data frame that contains the subject IDs. Default is 1. If the subject IDs are the row names of the original data frame, change this to 0.}
}
\value{
A data frame with two columns: one containing subject IDs, the other containing the percent missing for each subject.
}
\description{
Computes the percentage of missing items in a subscale for each subject.
}

