% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/centerer.R
\name{centerer}
\alias{centerer}
\title{Centers longitudinal measurements within and between subjects, as described by Bolger and Laurenceau (2013).}
\usage{
centerer(dataframe, measures, format, idxOfSubj = 1)
}
\arguments{
\item{dataframe}{A data frame containing the uncentered variables. May be wide or long.}

\item{measures}{A vector of the measures/subscales to be centered. FOR WIDE FORMAT, must be a character vector where the names uniquely match the appropriate column names of the data frame. For example, if the data were in wide format and mood was measured on multiple occasions labeled "Mood1", "Mood2," "Mood3," etc., the corresponding string could be "Mood." But if there were two different measures with mood in the title, like "MoodHusband1," "MoodHusband2," "MoodWife1", "MoodWife2", etc., the argument "Mood" would be insufficient because it would match both of these measures; you would need to include "MoodHusband" and "MoodWife" as separate strings in the vector. FOR LONG FORMAT, can be a vector of indices of the columns containing the measures/subscales or a character vector in which names EXACTLY match the column names.}

\item{format}{The format of the input data frame, either "long" or "wide".}

\item{idxOfSubj}{The column index in the data frame that contains subject IDs. Default is 1.}
}
\value{
A data frame containing the original data as well as the requested measures grand-mean centered, between-person centered, and within-person centered. Measures labeled "gmc" are grand-mean centered, measures labeled "cb" are between-persons centered (i.e., are the average value of the grand-mean centered score for each subject), and measures labeled "cw" are within-person centered (i.e., a subject's deviation from that person's own average).
}
\description{
Centers longitudinal measurements within and between subjects, as described by Bolger and Laurenceau (2013).
}

