% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/megaMissingness.R
\name{megaMissingness}
\alias{megaMissingness}
\title{Creates a data frame containing the percentage of items that are missing for a given set of subscales and measurement occasions for each subject.}
\usage{
megaMissingness(..., subscales, lookupList, idxOfSubj = 1,
  writeToExcel = FALSE, path = getwd())
}
\arguments{
\item{...}{One or more lists that contain named data frames as their elements. Each list in my data set corresponds to a measurement occasion.}

\item{subscales}{A character vector of the names of subscales for which the user wants to know missingness.}

\item{lookupList}{A list object that contains scoring "instructions" for each subscale.}

\item{idxOfSubj}{The column index in the data frames that contains subject IDs. Default is 1.}

\item{writeToExcel}{A logical vector of length one that specifies whether the resulting data frame should be written to a new Excel file. Default is FALSE. The file will be created in the current working directory and named with occassion and subscale names and the word "scored."}

\item{path}{The desired location at which to save the file. Default is the working directory.}
}
\value{
A data frame containing the percentage of values that are missing for each subject and occasion. That is, each subject gets a row, and each subscale-occasion combo gets its own column.
}
\description{
Creates a data frame containing the percentage of items that are missing for a given set of subscales and measurement occasions for each subject.
}

