% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/megaScorer.R
\name{megaScorer}
\alias{megaScorer}
\title{Scores requested subscales from a given set of lists of data frames.}
\usage{
megaScorer(filenames, subscales, lookupList, idxOfSubj = 1, myStartRow = 3,
  writeToExcel = FALSE, path = getwd(), meanSub = FALSE, criterion = 50)
}
\arguments{
\item{filenames}{A list of Excel files that contain named spreadsheets.}

\item{subscales}{A character vector of the names of subscales that the user wants to compute, written so the names match those in the lookup list.}

\item{lookupList}{A list object that contains scoring "instructions" for each subscale.}

\item{idxOfSubj}{The column index in the spreadsheets that contains subject IDs. Default is 1.}

\item{myStartRow}{The row index in the spreadsheets containing the variable names (i.e., headings). Defaults to row 3 because that's what I need for this project!}

\item{writeToExcel}{A logical vector of length one that specifies whether the resulting data frame should be written to a new Excel file. Default is FALSE. The file will be created in the current working directory and named with occassion and subscale names and the word "scored."}

\item{path}{The desired location at which to save the file. Default is the working directory.}

\item{meanSub}{Should item means be substituted before calculating the total? Default is FALSE.}

\item{criterion}{A number indicating the maximum percent of items a subject can have missing in order to use mean substitution. Default is 50, but a warning is produced if the value is over 20 because there are more robust ways to deal with missing values than mean substitution!}
}
\value{
A data frame containing the requested subscale scores for each subject and occasion. That is, each subject gets a row, and each subscale-occasion combo gets its own column.
}
\description{
Scores requested subscales from a given set of lists of data frames.
}

