% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/percentMissing.R
\name{percentMissing}
\alias{percentMissing}
\title{Computes the percentage of items that are missing for a given subject on a given subscale.}
\usage{
percentMissing(itemvec)
}
\arguments{
\item{itemvec}{A vector containing the items for which missingness is to be computed.}
}
\value{
The percent of items on the scale that are missing for that subject, rounded to two decimal places. 100 indicates that all items are missing, 0 indicates that no items are missing.
}
\description{
Computes the percentage of items that are missing for a given subject on a given subscale.
}

