% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/workbookToDfList.R
\name{workbookToDfList}
\alias{workbookToDfList}
\title{Read in an Excel workbook and produce a list of data frames, one for each sheet.}
\usage{
workbookToDfList(filename, mystartrow = 3)
}
\arguments{
\item{filename}{The name/path of an Excel file (xls or xlsx)}
}
\value{
A list of data frames, one for each sheet, with the name of the list element being the name of the sheet.
}
\description{
Read in an Excel workbook and produce a list of data frames, one for each sheet.
}

