% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLookup.R
\name{getLookup}
\alias{getLookup}
\title{Reads in and cleans up the scoring lookup list.}
\usage{
getLookup(filename, mystartrow = 1)
}
\arguments{
\item{filename}{The location of the file containing the scoring lookup list. This function assumes that it is an Excel file formatted in a very specific way.}
}
\value{
A list of lists, one list for each subscale, named with the subscale name and containing four sublists: measName, forwItems, revItems, and revInt.
}
\description{
Reads in and cleans up the scoring lookup list.
}

